mini_unset is upset, original **env might need to be duplicated, mini_env should display **env not env_list
