+++ Makefile
@@ -35,6 +35,8 @@ SRCS 	:=	src/main.c \
 			src/builtin/mini_pwd.c \
 			src/builtin/mini_env.c \
 			src/builtin/mini_exit.c \
+			src/builtin/mini_unset.c \
+			src/builtin/builtin_utils.c \
 			src/utils/utils.c
 
 #SRCS	:= ${SRCS:%=./src/%}
